test-requirements.txt
Faker==13.3.4
pylint==2.13.5
pylint-django==2.5.3
